+++ irs-api/src/main/java/org/eclipse/tractusx/irs/connector/job/MultiTransferJob.java
@@ -97,7 +97,7 @@ public class MultiTransferJob {
 
     @JsonIgnore
     public JobParameter getJobParameter() {
-        return getJob().getJobParameter();
+        return getJob().getParameter();
     }
 
     /**