+++ irs-api/src/test/java/org/eclipse/tractusx/irs/connector/job/InMemoryJobStoreTest.java
@@ -411,7 +411,7 @@ class InMemoryJobStoreTest {
                                                                  .exception("SomeError")
                                                                  .exceptionDate(ZonedDateTime.now())
                                                                  .build())
-                                       .jobParameter(jobParameter())
+                                       .parameter(jobParameter())
                                        .build())
                                .build();
     }